+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.5.1 - Not yet released
 - Fixes not able to create multiple PDF documents per request
 - Updates documentation
+- Fixes PHP 5.4 compatibility error
 
 ## 1.5.0 - May 4, 2018
 - Updates documentation